Split ctl/command.rs into modules
* Create new modules in addition to the command one: `request_builder` and `display`
