+++ bin/src/ctl/command.rs
@@ -1,30 +1,31 @@
-use crate::cli::{LoggingLevel, MetricsCmd};
 use sozu_command_lib::{
-    certificate::{calculate_fingerprint, split_certificate_chain},
     command::{
         CommandRequest, CommandRequestData, CommandResponse, CommandResponseData, CommandStatus,
         FrontendFilters, RunState, WorkerInfo,
     },
-    config::{Config, FileListenerProtocolConfig, Listener, ProxyProtocolConfig},
     proxy::{
-        ActivateListener, AddCertificate, Backend, CertificateAndKey, CertificateFingerprint,
-        Cluster, DeactivateListener, FilteredData, HttpFrontend, ListenerType,
-        LoadBalancingAlgorithms, LoadBalancingParams, MetricsConfiguration, PathRule,
-        ProxyRequestData, Query, QueryAnswer, QueryAnswerCertificate, QueryAnswerMetrics,
-        QueryApplicationDomain, QueryApplicationType, QueryCertificateType, QueryMetricsType,
-        RemoveBackend, RemoveCertificate, RemoveListener, ReplaceCertificate, Route, RulePosition,
-        TcpFrontend, TcpListener, TlsVersion,
+        MetricsConfiguration, ProxyRequestData, Query, QueryApplicationDomain,
+        QueryApplicationType, QueryCertificateType, QueryMetricsType,
     },
 };
 
-use super::{create_channel, CommandManager};
+use crate::{
+    cli::MetricsCmd,
+    ctl::{
+        create_channel,
+        display::{
+            print_certificates, print_frontend_list, print_json_response, print_metrics,
+            print_query_response_data,
+        },
+        CommandManager,
+    },
+};
+
 use anyhow::{self, bail, Context};
-use prettytable::{Row, Table};
+use prettytable::Table;
 use rand::{distributions::Alphanumeric, thread_rng, Rng};
-use serde_json;
 use std::{
-    collections::{BTreeMap, HashMap, HashSet},
-    net::SocketAddr,
+    collections::{HashMap, HashSet},
     process::exit,
     sync::mpsc,
     sync::{Arc, Mutex},
@@ -224,6 +225,9 @@ impl CommandManager {
         Ok(())
     }
 
+    // 1. Request a list of workers
+    // 2. Send an UpgradeMain
+    // 3. Send an UpgradeWorker to each worker
     pub fn upgrade_main(&mut self) -> Result<(), anyhow::Error> {
         println!("Preparing to upgrade proxy...");
 
@@ -612,253 +616,6 @@ impl CommandManager {
         Ok(())
     }
 
-    pub fn add_application(
-        &mut self,
-        cluster_id: &str,
-        sticky_session: bool,
-        https_redirect: bool,
-        send_proxy: bool,
-        expect_proxy: bool,
-        load_balancing: LoadBalancingAlgorithms,
-    ) -> Result<(), anyhow::Error> {
-        let proxy_protocol = match (send_proxy, expect_proxy) {
-            (true, true) => Some(ProxyProtocolConfig::RelayHeader),
-            (true, false) => Some(ProxyProtocolConfig::SendHeader),
-            (false, true) => Some(ProxyProtocolConfig::ExpectHeader),
-            _ => None,
-        };
-
-        self.order_command(ProxyRequestData::AddCluster(Cluster {
-            cluster_id: String::from(cluster_id),
-            sticky_session,
-            https_redirect,
-            proxy_protocol,
-            load_balancing,
-            load_metric: None,
-            answer_503: None,
-        }))
-    }
-
-    pub fn remove_application(&mut self, cluster_id: &str) -> Result<(), anyhow::Error> {
-        self.order_command(ProxyRequestData::RemoveCluster {
-            cluster_id: String::from(cluster_id),
-        })
-    }
-
-    pub fn add_http_frontend(
-        &mut self,
-        route: Route,
-        address: SocketAddr,
-        hostname: &str,
-        path: &str,
-        method: Option<&str>,
-        https: bool,
-    ) -> Result<(), anyhow::Error> {
-        if https {
-            self.order_command(ProxyRequestData::AddHttpsFrontend(HttpFrontend {
-                route,
-                address,
-                hostname: String::from(hostname),
-                path: PathRule::Prefix(String::from(path)),
-                method: method.map(String::from),
-                position: RulePosition::Tree,
-            }))
-        } else {
-            self.order_command(ProxyRequestData::AddHttpFrontend(HttpFrontend {
-                route,
-                address,
-                hostname: String::from(hostname),
-                path: PathRule::Prefix(String::from(path)),
-                method: method.map(String::from),
-                position: RulePosition::Tree,
-            }))
-        }
-    }
-
-    pub fn remove_http_frontend(
-        &mut self,
-        route: Route,
-        address: SocketAddr,
-        hostname: &str,
-        path: &str,
-        method: Option<&str>,
-        https: bool,
-    ) -> Result<(), anyhow::Error> {
-        if https {
-            self.order_command(ProxyRequestData::RemoveHttpsFrontend(HttpFrontend {
-                route,
-                address,
-                hostname: String::from(hostname),
-                path: PathRule::Prefix(String::from(path)),
-                method: method.map(String::from),
-                position: RulePosition::Tree,
-            }))
-        } else {
-            self.order_command(ProxyRequestData::RemoveHttpFrontend(HttpFrontend {
-                route,
-                address,
-                hostname: String::from(hostname),
-                path: PathRule::Prefix(String::from(path)),
-                method: method.map(String::from),
-                position: RulePosition::Tree,
-            }))
-        }
-    }
-
-    pub fn add_backend(
-        &mut self,
-        cluster_id: &str,
-        backend_id: &str,
-        address: SocketAddr,
-        sticky_id: Option<String>,
-        backup: Option<bool>,
-    ) -> Result<(), anyhow::Error> {
-        self.order_command(ProxyRequestData::AddBackend(Backend {
-            cluster_id: String::from(cluster_id),
-            address: address,
-            backend_id: String::from(backend_id),
-            load_balancing_parameters: Some(LoadBalancingParams::default()),
-            sticky_id: sticky_id,
-            backup: backup,
-        }))
-    }
-
-    pub fn remove_backend(
-        &mut self,
-        cluster_id: &str,
-        backend_id: &str,
-        address: SocketAddr,
-    ) -> Result<(), anyhow::Error> {
-        self.order_command(ProxyRequestData::RemoveBackend(RemoveBackend {
-            cluster_id: String::from(cluster_id),
-            address: address,
-            backend_id: String::from(backend_id),
-        }))
-    }
-
-    pub fn add_certificate(
-        &mut self,
-        address: SocketAddr,
-        certificate_path: &str,
-        certificate_chain_path: &str,
-        key_path: &str,
-        versions: Vec<TlsVersion>,
-    ) -> Result<(), anyhow::Error> {
-        if let Some(new_certificate) =
-            load_full_certificate(certificate_path, certificate_chain_path, key_path, versions)?
-        {
-            self.order_command(ProxyRequestData::AddCertificate(AddCertificate {
-                address,
-                certificate: new_certificate,
-                names: vec![],
-                expired_at: None,
-            }))?;
-        }
-        Ok(())
-    }
-
-    pub fn remove_certificate(
-        &mut self,
-        address: SocketAddr,
-        certificate_path: Option<&str>,
-        fingerprint: Option<&str>,
-    ) -> Result<(), anyhow::Error> {
-        if certificate_path.is_some() && fingerprint.is_some() {
-            bail!("Error: Either provide the certificate's path or its fingerprint");
-        }
-
-        if certificate_path.is_none() && fingerprint.is_none() {
-            bail!("Error: Either provide the certificate's path or its fingerprint");
-        }
-
-        if let Some(fingerprint) = fingerprint
-            .and_then(|s| match hex::decode(s) {
-                Ok(v) => Some(CertificateFingerprint(v)),
-                Err(e) => {
-                    eprintln!(
-                    "Error decoding the certificate fingerprint (expected hexadecimal data): {:?}",
-                    e
-                );
-                    None
-                }
-            })
-            .or(certificate_path.and_then(get_certificate_fingerprint))
-        {
-            self.order_command(ProxyRequestData::RemoveCertificate(RemoveCertificate {
-                address,
-                fingerprint,
-            }))?
-        }
-        Ok(())
-    }
-
-    pub fn replace_certificate(
-        &mut self,
-        address: SocketAddr,
-        new_certificate_path: &str,
-        new_certificate_chain_path: &str,
-        new_key_path: &str,
-        old_certificate_path: Option<&str>,
-        old_fingerprint: Option<&str>,
-        versions: Vec<TlsVersion>,
-    ) -> Result<(), anyhow::Error> {
-        if old_certificate_path.is_some() && old_fingerprint.is_some() {
-            bail!("Error: Either provide the old certificate's path or its fingerprint");
-        }
-
-        if old_certificate_path.is_none() && old_fingerprint.is_none() {
-            bail!("Error: Either provide the old certificate's path or its fingerprint");
-        }
-
-        if let Some(new_certificate) = load_full_certificate(
-            new_certificate_path,
-            new_certificate_chain_path,
-            new_key_path,
-            versions,
-        )? {
-            if let Some(old_fingerprint) = old_fingerprint.and_then(|s| {
-        match hex::decode(s) {
-            Ok(v) => Some(CertificateFingerprint(v)),
-            Err(e) => {
-                eprintln!("Error decoding the certificate fingerprint (expected hexadecimal data): {:?}", e);
-                None
-            }
-        }
-    }).or(old_certificate_path.and_then(get_certificate_fingerprint)) {
-     self.order_command( ProxyRequestData::ReplaceCertificate(ReplaceCertificate {
-        address,
-        new_certificate,
-        old_fingerprint,
-        new_names: vec![],
-        new_expired_at: None,
-      }))?;
-    }
-        }
-        Ok(())
-    }
-
-    pub fn add_tcp_frontend(
-        &mut self,
-        cluster_id: &str,
-        address: SocketAddr,
-    ) -> Result<(), anyhow::Error> {
-        self.order_command(ProxyRequestData::AddTcpFrontend(TcpFrontend {
-            cluster_id: String::from(cluster_id),
-            address,
-        }))
-    }
-
-    pub fn remove_tcp_frontend(
-        &mut self,
-        cluster_id: &str,
-        address: SocketAddr,
-    ) -> Result<(), anyhow::Error> {
-        self.order_command(ProxyRequestData::RemoveTcpFrontend(TcpFrontend {
-            cluster_id: String::from(cluster_id),
-            address,
-        }))
-    }
-
     pub fn list_frontends(
         &mut self,
         http: bool,
@@ -890,176 +647,13 @@ impl CommandManager {
             CommandStatus::Ok => {
                 debug!("We received this response: {:?}", message.data);
 
-                if let Some(CommandResponseData::FrontendList(data)) = message.data {
-                    trace!(" We received this data to display {:#?}", data);
-                    // HTTP frontends
-                    if !data.http_frontends.is_empty() {
-                        let mut table = Table::new();
-                        table.add_row(row!["HTTP frontends"]);
-                        table.add_row(row![
-                            "route", "address", "hostname", "path", "method", "position"
-                        ]);
-                        for http_frontend in data.http_frontends.iter() {
-                            table.add_row(row!(
-                                http_frontend.route,
-                                http_frontend.address.to_string(),
-                                http_frontend.hostname.to_string(),
-                                format!("{:?}", http_frontend.path),
-                                format!("{:?}", http_frontend.method),
-                                format!("{:?}", http_frontend.position),
-                            ));
-                        }
-                        table.printstd();
-                    }
-
-                    // HTTPS frontends
-                    if !data.https_frontends.is_empty() {
-                        let mut table = Table::new();
-                        table.add_row(row!["HTTPS frontends"]);
-                        table.add_row(row![
-                            "route", "address", "hostname", "path", "method", "position"
-                        ]);
-                        for https_frontend in data.https_frontends.iter() {
-                            table.add_row(row!(
-                                https_frontend.route,
-                                https_frontend.address.to_string(),
-                                https_frontend.hostname.to_string(),
-                                format!("{:?}", https_frontend.path),
-                                format!("{:?}", https_frontend.method),
-                                format!("{:?}", https_frontend.position),
-                            ));
-                        }
-                        table.printstd();
-                    }
-
-                    // TCP frontends
-                    if !data.tcp_frontends.is_empty() {
-                        let mut table = Table::new();
-                        table.add_row(row!["TCP frontends"]);
-                        table.add_row(row!["Cluster ID", "address"]);
-                        for tcp_frontend in data.tcp_frontends.iter() {
-                            table.add_row(row!(tcp_frontend.cluster_id, tcp_frontend.address,));
-                        }
-                        table.printstd();
-                    }
+                if let Some(CommandResponseData::FrontendList(frontends)) = message.data {
+                    print_frontend_list(frontends);
                 }
             }
         }
 
         Ok(())
-    }
-
-    pub fn add_http_listener(
-        &mut self,
-        address: SocketAddr,
-        public_address: Option<SocketAddr>,
-        answer_404: Option<String>,
-        answer_503: Option<String>,
-        expect_proxy: bool,
-        sticky_name: Option<String>,
-    ) -> Result<(), anyhow::Error> {
-        let mut listener = Listener::new(address, FileListenerProtocolConfig::Http);
-        listener.public_address = public_address;
-        listener.answer_404 = answer_404;
-        listener.answer_503 = answer_503;
-        listener.expect_proxy = Some(expect_proxy);
-        if let Some(sticky_name) = sticky_name {
-            listener.sticky_name = sticky_name;
-        }
-
-        match listener.to_http(None, None, None) {
-            Some(conf) => self.order_command(ProxyRequestData::AddHttpListener(conf)),
-            None => bail!("Error creating HTTP listener"),
-        }
-    }
-
-    pub fn add_https_listener(
-        &mut self,
-        address: SocketAddr,
-        public_address: Option<SocketAddr>,
-        answer_404: Option<String>,
-        answer_503: Option<String>,
-        tls_versions: Vec<TlsVersion>,
-        cipher_list: Option<String>,
-        rustls_cipher_list: Vec<String>,
-        expect_proxy: bool,
-        sticky_name: Option<String>,
-    ) -> Result<(), anyhow::Error> {
-        let mut listener = Listener::new(address, FileListenerProtocolConfig::Https);
-        listener.public_address = public_address;
-        listener.answer_404 = answer_404;
-        listener.answer_503 = answer_503;
-        listener.expect_proxy = Some(expect_proxy);
-        if let Some(sticky_name) = sticky_name {
-            listener.sticky_name = sticky_name;
-        }
-        listener.cipher_list = cipher_list;
-        listener.tls_versions = if tls_versions.len() == 0 {
-            None
-        } else {
-            Some(tls_versions)
-        };
-        listener.rustls_cipher_list = if rustls_cipher_list.len() == 0 {
-            None
-        } else {
-            Some(rustls_cipher_list)
-        };
-
-        match listener.to_tls(None, None, None) {
-            Some(conf) => self.order_command(ProxyRequestData::AddHttpsListener(conf)),
-            None => bail!("Error creating HTTPS listener"),
-        }
-    }
-
-    pub fn add_tcp_listener(
-        &mut self,
-        address: SocketAddr,
-        public_address: Option<SocketAddr>,
-        expect_proxy: bool,
-    ) -> Result<(), anyhow::Error> {
-        self.order_command(ProxyRequestData::AddTcpListener(TcpListener {
-            address,
-            public_address,
-            expect_proxy,
-            front_timeout: 60,
-            back_timeout: 30,
-            connect_timeout: 3,
-        }))
-    }
-
-    pub fn remove_listener(
-        &mut self,
-        address: SocketAddr,
-        proxy: ListenerType,
-    ) -> Result<(), anyhow::Error> {
-        self.order_command(ProxyRequestData::RemoveListener(RemoveListener {
-            address,
-            proxy,
-        }))
-    }
-
-    pub fn activate_listener(
-        &mut self,
-        address: SocketAddr,
-        proxy: ListenerType,
-    ) -> Result<(), anyhow::Error> {
-        self.order_command(ProxyRequestData::ActivateListener(ActivateListener {
-            address,
-            proxy,
-            from_scm: false,
-        }))
-    }
-
-    pub fn deactivate_listener(
-        &mut self,
-        address: SocketAddr,
-        proxy: ListenerType,
-    ) -> Result<(), anyhow::Error> {
-        self.order_command(ProxyRequestData::DeactivateListener(DeactivateListener {
-            address,
-            proxy,
-            to_scm: false,
-        }))
     }
 
     pub fn query_application(
@@ -1124,246 +718,7 @@ impl CommandManager {
                 bail!("could not query proxy state: {}", message.message);
             }
             CommandStatus::Ok => {
-                if let Some(needle) = application_id.or(domain) {
-                    if let Some(CommandResponseData::Query(data)) = message.data {
-                        if json {
-                            return Ok(print_json_response(&data)?);
-                        }
-
-                        let application_headers = vec!["id", "sticky_session", "https_redirect"];
-                        let mut application_table =
-                            create_queried_application_table(application_headers, &data);
-
-                        let http_headers = vec!["id", "hostname", "path"];
-                        let mut frontend_table =
-                            create_queried_application_table(http_headers, &data);
-
-                        let https_headers = vec!["id", "hostname", "path"];
-                        let mut https_frontend_table =
-                            create_queried_application_table(https_headers, &data);
-
-                        let tcp_headers = vec!["id", "address"];
-                        let mut tcp_frontend_table =
-                            create_queried_application_table(tcp_headers, &data);
-
-                        let backend_headers = vec!["backend id", "IP address", "Backup"];
-                        let mut backend_table =
-                            create_queried_application_table(backend_headers, &data);
-
-                        let keys: HashSet<&String> = data.keys().collect();
-
-                        let mut application_data = HashMap::new();
-                        let mut frontend_data = HashMap::new();
-                        let mut https_frontend_data = HashMap::new();
-                        let mut tcp_frontend_data = HashMap::new();
-                        let mut backend_data = HashMap::new();
-
-                        for (ref key, ref metrics) in data.iter() {
-                            //let m: u8 = metrics;
-                            if let &QueryAnswer::Applications(ref apps) = *metrics {
-                                for app in apps.iter() {
-                                    let entry = application_data.entry(app).or_insert(Vec::new());
-                                    entry.push((*key).clone());
-
-                                    for frontend in app.http_frontends.iter() {
-                                        let entry =
-                                            frontend_data.entry(frontend).or_insert(Vec::new());
-                                        entry.push((*key).clone());
-                                    }
-
-                                    for frontend in app.https_frontends.iter() {
-                                        let entry = https_frontend_data
-                                            .entry(frontend)
-                                            .or_insert(Vec::new());
-                                        entry.push((*key).clone());
-                                    }
-
-                                    for frontend in app.tcp_frontends.iter() {
-                                        let entry =
-                                            tcp_frontend_data.entry(frontend).or_insert(Vec::new());
-                                        entry.push((*key).clone());
-                                    }
-
-                                    for backend in app.backends.iter() {
-                                        let entry =
-                                            backend_data.entry(backend).or_insert(Vec::new());
-                                        entry.push((*key).clone());
-                                    }
-                                }
-                            }
-                        }
-
-                        println!("Cluster level configuration for {}:\n", needle);
-
-                        for (ref key, ref values) in application_data.iter() {
-                            let mut row = Vec::new();
-                            row.push(cell!(key
-                                .configuration
-                                .clone()
-                                .map(|conf| conf.cluster_id)
-                                .unwrap_or(String::from(""))));
-                            row.push(cell!(key
-                                .configuration
-                                .clone()
-                                .map(|conf| conf.sticky_session)
-                                .unwrap_or(false)));
-                            row.push(cell!(key
-                                .configuration
-                                .clone()
-                                .map(|conf| conf.https_redirect)
-                                .unwrap_or(false)));
-
-                            for val in values.iter() {
-                                if keys.contains(val) {
-                                    row.push(cell!(String::from("X")));
-                                } else {
-                                    row.push(cell!(String::from("")));
-                                }
-                            }
-
-                            application_table.add_row(Row::new(row));
-                        }
-
-                        application_table.printstd();
-
-                        println!("\nHTTP frontends configuration for {}:\n", needle);
-
-                        for (ref key, ref values) in frontend_data.iter() {
-                            let mut row = Vec::new();
-                            match &key.route {
-                                Route::ClusterId(cluster_id) => row.push(cell!(cluster_id)),
-                                Route::Deny => row.push(cell!("-")),
-                            }
-                            row.push(cell!(key.hostname));
-                            row.push(cell!(key.path));
-
-                            for val in values.iter() {
-                                if keys.contains(val) {
-                                    row.push(cell!(String::from("X")));
-                                } else {
-                                    row.push(cell!(String::from("")));
-                                }
-                            }
-
-                            frontend_table.add_row(Row::new(row));
-                        }
-
-                        frontend_table.printstd();
-
-                        println!("\nHTTPS frontends configuration for {}:\n", needle);
-
-                        for (ref key, ref values) in https_frontend_data.iter() {
-                            let mut row = Vec::new();
-                            match &key.route {
-                                Route::ClusterId(cluster_id) => row.push(cell!(cluster_id)),
-                                Route::Deny => row.push(cell!("-")),
-                            }
-                            row.push(cell!(key.hostname));
-                            row.push(cell!(key.path));
-
-                            for val in values.iter() {
-                                if keys.contains(val) {
-                                    row.push(cell!(String::from("X")));
-                                } else {
-                                    row.push(cell!(String::from("")));
-                                }
-                            }
-
-                            https_frontend_table.add_row(Row::new(row));
-                        }
-
-                        https_frontend_table.printstd();
-
-                        println!("\nTCP frontends configuration for {}:\n", needle);
-
-                        for (ref key, ref values) in tcp_frontend_data.iter() {
-                            let mut row = Vec::new();
-                            row.push(cell!(key.cluster_id));
-                            row.push(cell!(format!("{}", key.address)));
-
-                            for val in values.iter() {
-                                if keys.contains(val) {
-                                    row.push(cell!(String::from("X")));
-                                } else {
-                                    row.push(cell!(String::from("")));
-                                }
-                            }
-
-                            tcp_frontend_table.add_row(Row::new(row));
-                        }
-
-                        tcp_frontend_table.printstd();
-
-                        println!("\nbackends configuration for {}:\n", needle);
-
-                        for (ref key, ref values) in backend_data.iter() {
-                            let mut row = Vec::new();
-                            let backend_backup =
-                                key.backup.map(|b| if b { "X" } else { "" }).unwrap_or("");
-                            row.push(cell!(key.backend_id));
-                            row.push(cell!(format!("{}", key.address)));
-                            row.push(cell!(backend_backup));
-
-                            for val in values.iter() {
-                                if keys.contains(val) {
-                                    row.push(cell!(String::from("X")));
-                                } else {
-                                    row.push(cell!(String::from("")));
-                                }
-                            }
-
-                            backend_table.add_row(Row::new(row));
-                        }
-
-                        backend_table.printstd();
-                    }
-                } else {
-                    if let Some(CommandResponseData::Query(data)) = message.data {
-                        let mut table = Table::new();
-                        let mut header = Vec::new();
-                        header.push(cell!("key"));
-                        for ref key in data.keys() {
-                            header.push(cell!(&key));
-                        }
-                        header.push(cell!("desynchronized"));
-                        table.add_row(Row::new(header));
-
-                        let mut query_data = HashMap::new();
-
-                        for ref metrics in data.values() {
-                            //let m: u8 = metrics;
-                            if let &QueryAnswer::ApplicationsHashes(ref apps) = *metrics {
-                                for (ref key, ref value) in apps.iter() {
-                                    (*(query_data.entry((*key).clone()).or_insert(Vec::new())))
-                                        .push(*value);
-                                }
-                            }
-                        }
-
-                        for (ref key, ref values) in query_data.iter() {
-                            let mut row = Vec::new();
-                            row.push(cell!(key));
-
-                            for val in values.iter() {
-                                row.push(cell!(format!("{}", val)));
-                            }
-
-                            let hs: HashSet<&u64> = values.iter().cloned().collect();
-
-                            let diff = hs.len() > 1;
-
-                            if diff {
-                                row.push(cell!(String::from("X")));
-                            } else {
-                                row.push(cell!(String::from("")));
-                            }
-
-                            table.add_row(Row::new(row));
-                        }
-
-                        table.printstd();
-                    }
-                }
+                print_query_response_data(application_id, domain, message.data, json)?;
             }
         }
 
@@ -1418,57 +773,7 @@ impl CommandManager {
             }
             CommandStatus::Ok => {
                 if let Some(CommandResponseData::Query(data)) = message.data {
-                    if json {
-                        print_json_response(&data)?;
-                        return Ok(());
-                    }
-
-                    //println!("received: {:?}", data);
-                    let it = data.iter().map(|(k, v)| match v {
-                        QueryAnswer::Certificates(c) => (k, c),
-                        v => {
-                            eprintln!("unexpected certificates query answer: {:?}", v);
-                            exit(1);
-                        }
-                    });
-
-                    for (k, v) in it {
-                        println!("process '{}':", k);
-
-                        match v {
-                            QueryAnswerCertificate::All(h) => {
-                                for (addr, h2) in h.iter() {
-                                    println!("\t{}:", addr);
-
-                                    for (domain, fingerprint) in h2.iter() {
-                                        println!("\t\t{}:\t{}", domain, hex::encode(fingerprint));
-                                    }
-
-                                    println!("");
-                                }
-                            }
-                            QueryAnswerCertificate::Domain(h) => {
-                                for (addr, opt) in h.iter() {
-                                    println!("\t{}:", addr);
-                                    if let Some((key, fingerprint)) = opt {
-                                        println!("\t\t{}:\t{}", key, hex::encode(fingerprint));
-                                    } else {
-                                        println!("\t\tnot found");
-                                    }
-
-                                    println!("");
-                                }
-                            }
-                            QueryAnswerCertificate::Fingerprint(opt) => {
-                                if let Some((s, v)) = opt {
-                                    println!("\tfrontends: {:?}\ncertificate:\n{}", v, s);
-                                } else {
-                                    println!("\tnot found");
-                                }
-                            }
-                        }
-                        println!("");
-                    }
+                    print_certificates(data, json)?;
                 } else {
                     bail!("unexpected response: {:?}", message.data);
                 }
@@ -1536,73 +841,8 @@ impl CommandManager {
                     }
                 }
                 CommandStatus::Ok => {
-                    if let Some(CommandResponseData::Query(data)) = message.data {
-                        if json {
-                            return print_json_response(&data);
-                        }
-
-                        //println!("got data: {:#?}", data);
-                        if list {
-                            let metrics: HashSet<_> = data
-                                .values()
-                                .filter_map(|value| match value {
-                                    QueryAnswer::Metrics(QueryAnswerMetrics::List(v)) => {
-                                        Some(v.iter())
-                                    }
-                                    _ => None,
-                                })
-                                .flatten()
-                                .map(|s| s.replace("\t", "."))
-                                .collect();
-                            let mut metrics: Vec<_> = metrics.iter().collect();
-                            metrics.sort();
-                            println!("available metrics: {:?}", metrics);
-                            return Ok(());
-                        }
-
-                        let data = data
-                            .iter()
-                            .filter_map(|(key, value)| match value {
-                                QueryAnswer::Metrics(QueryAnswerMetrics::Cluster(d)) => {
-                                    let mut metrics = BTreeMap::new();
-                                    for (cluster_id, cluster_metrics) in d.iter() {
-                                        for (metric_key, value) in cluster_metrics.iter() {
-                                            metrics.insert(
-                                                format!(
-                                                    "{} {}",
-                                                    cluster_id,
-                                                    metric_key.replace("\t", ".")
-                                                ),
-                                                value.clone(),
-                                            );
-                                        }
-                                    }
-                                    Some((key.clone(), metrics))
-                                }
-                                QueryAnswer::Metrics(QueryAnswerMetrics::Backend(d)) => {
-                                    let mut metrics = BTreeMap::new();
-                                    for (cluster_id, cluster_metrics) in d.iter() {
-                                        for (backend_id, backend_metrics) in cluster_metrics.iter()
-                                        {
-                                            for (metric_key, value) in backend_metrics.iter() {
-                                                metrics.insert(
-                                                    format!(
-                                                        "{}/{} {}",
-                                                        cluster_id,
-                                                        backend_id,
-                                                        metric_key.replace("\t", ".")
-                                                    ),
-                                                    value.clone(),
-                                                );
-                                            }
-                                        }
-                                    }
-                                    Some((key.clone(), metrics))
-                                }
-                                _ => None,
-                            })
-                            .collect::<BTreeMap<_, _>>();
-                        print_metrics("Result", &data);
+                    if let Some(CommandResponseData::Query(answers)) = message.data {
+                        print_metrics(answers, json, list)?;
                     }
                 }
             }
@@ -1620,10 +860,6 @@ impl CommandManager {
         }
 
         Ok(())
-    }
-
-    pub fn logging_filter(&mut self, filter: &LoggingLevel) -> Result<(), anyhow::Error> {
-        self.order_command(ProxyRequestData::Logging(filter.to_string().to_lowercase()))
     }
 
     pub fn events(&mut self) -> Result<(), anyhow::Error> {
@@ -1651,7 +887,7 @@ impl CommandManager {
         Ok(())
     }
 
-    fn order_command(&mut self, order: ProxyRequestData) -> Result<(), anyhow::Error> {
+    pub fn order_command(&mut self, order: ProxyRequestData) -> Result<(), anyhow::Error> {
         let id = generate_id();
         self.channel.write_message(&CommandRequest::new(
             id.clone(),
@@ -1692,187 +928,4 @@ impl CommandManager {
         }
         Ok(())
     }
-}
-
-fn print_json_response<T: ::serde::Serialize>(input: &T) -> Result<(), anyhow::Error> {
-    println!(
-        "{}",
-        serde_json::to_string_pretty(&input).context("Error while parsing response to JSON")?
-    );
-    Ok(())
-}
-
-fn load_full_certificate(
-    certificate_path: &str,
-    certificate_chain_path: &str,
-    key_path: &str,
-    versions: Vec<TlsVersion>,
-) -> Result<Option<CertificateAndKey>, anyhow::Error> {
-    match Config::load_file(certificate_path) {
-        Err(e) => {
-            bail!("could not load certificate: {:?}", e);
-        }
-        Ok(certificate) => {
-            match Config::load_file(certificate_chain_path).map(split_certificate_chain) {
-                Err(e) => {
-                    bail!("could not load certificate chain: {:?}", e);
-                }
-                Ok(certificate_chain) => match Config::load_file(key_path) {
-                    Err(e) => {
-                        bail!("could not load key: {:?}", e);
-                    }
-                    Ok(key) => Ok(Some(CertificateAndKey {
-                        certificate,
-                        certificate_chain,
-                        key,
-                        versions,
-                    })),
-                },
-            }
-        }
-    }
-}
-
-// input: map worker_id -> (map key -> value)
-fn print_metrics(table_name: &str, data: &BTreeMap<String, BTreeMap<String, FilteredData>>) {
-    let mut metrics = data
-        .values()
-        .flat_map(|map| {
-            map.iter().filter_map(|(k, v)| match v {
-                FilteredData::Count(_) | FilteredData::Gauge(_) => Some(k),
-                _ => None,
-            })
-        })
-        .collect::<HashSet<_>>();
-
-    // sort the metrics so they always appear in the same order
-    let mut metrics: Vec<_> = metrics.drain().collect();
-    metrics.sort();
-    //println!("metrics list: {:?}", metrics);
-
-    if !metrics.is_empty() {
-        let mut table = Table::new();
-        table.set_format(*prettytable::format::consts::FORMAT_NO_BORDER_LINE_SEPARATOR);
-
-        let mut row = vec![cell!(table_name)];
-        for key in data.keys() {
-            row.push(cell!(key));
-        }
-        table.set_titles(Row::new(row));
-
-        for metric in metrics {
-            let mut row = vec![cell!(metric)];
-            for worker_data in data.values() {
-                match worker_data.get(metric) {
-                    Some(FilteredData::Count(c)) => row.push(cell!(c)),
-                    Some(FilteredData::Gauge(c)) => row.push(cell!(c)),
-                    _ => row.push(cell!("")),
-                }
-            }
-            table.add_row(Row::new(row));
-        }
-
-        table.printstd();
-    }
-
-    let mut time_metrics = data
-        .values()
-        .flat_map(|map| {
-            map.iter().filter_map(|(k, v)| match v {
-                FilteredData::Percentiles(_) => Some(k),
-                _ => None,
-            })
-        })
-        .collect::<HashSet<_>>();
-
-    // sort the metrics so they always appear in the same order
-    let mut time_metrics: Vec<_> = time_metrics.drain().collect();
-    time_metrics.sort();
-    //println!("time metrics list: {:?}", time_metrics);
-
-    if !time_metrics.is_empty() {
-        let mut timing_table = Table::new();
-        timing_table.set_format(*prettytable::format::consts::FORMAT_NO_BORDER_LINE_SEPARATOR);
-
-        let mut row = vec![cell!(table_name)];
-        for key in data.keys() {
-            row.push(cell!(key));
-        }
-        timing_table.set_titles(Row::new(row));
-
-        for metric in time_metrics {
-            let mut row_samples = vec![cell!(format!("{}.samples", metric))];
-            let mut row_p50 = vec![cell!(format!("{}.p50", metric))];
-            let mut row_p90 = vec![cell!(format!("{}.p90", metric))];
-            let mut row_p99 = vec![cell!(format!("{}.p99", metric))];
-            let mut row_p99_9 = vec![cell!(format!("{}.p99.9", metric))];
-            let mut row_p99_99 = vec![cell!(format!("{}.p99.99", metric))];
-            let mut row_p99_999 = vec![cell!(format!("{}.p99.999", metric))];
-            let mut row_p100 = vec![cell!(format!("{}.p100", metric))];
-
-            for worker_data in data.values() {
-                match worker_data.get(metric) {
-                    Some(FilteredData::Percentiles(p)) => {
-                        row_samples.push(cell!(p.samples));
-                        row_p50.push(cell!(p.p_50));
-                        row_p90.push(cell!(p.p_90));
-                        row_p99.push(cell!(p.p_99));
-                        row_p99_9.push(cell!(p.p_99_9));
-                        row_p99_99.push(cell!(p.p_99_99));
-                        row_p99_999.push(cell!(p.p_99_999));
-                        row_p100.push(cell!(p.p_100));
-                    }
-                    _ => {
-                        row_samples.push(cell!(""));
-                        row_p50.push(cell!(""));
-                        row_p90.push(cell!(""));
-                        row_p99.push(cell!(""));
-                        row_p99_9.push(cell!(""));
-                        row_p99_99.push(cell!(""));
-                        row_p99_999.push(cell!(""));
-                        row_p100.push(cell!(""));
-                    }
-                }
-            }
-            timing_table.add_row(Row::new(row_samples));
-            timing_table.add_row(Row::new(row_p50));
-            timing_table.add_row(Row::new(row_p90));
-            timing_table.add_row(Row::new(row_p99));
-            timing_table.add_row(Row::new(row_p99_9));
-            timing_table.add_row(Row::new(row_p99_99));
-            timing_table.add_row(Row::new(row_p99_999));
-            timing_table.add_row(Row::new(row_p100));
-        }
-
-        timing_table.printstd();
-    }
-}
-
-fn get_certificate_fingerprint(certificate_path: &str) -> Option<CertificateFingerprint> {
-    match Config::load_file_bytes(certificate_path) {
-        Ok(data) => match calculate_fingerprint(&data) {
-            Some(fingerprint) => Some(CertificateFingerprint(fingerprint)),
-            None => {
-                eprintln!("could not calculate finrprint for certificate");
-                exit(1);
-            }
-        },
-        Err(e) => {
-            eprintln!("could not load file: {:?}", e);
-            exit(1);
-        }
-    }
-}
-
-fn create_queried_application_table(
-    headers: Vec<&str>,
-    data: &BTreeMap<String, QueryAnswer>,
-) -> Table {
-    let mut table = Table::new();
-    let mut row_header: Vec<_> = headers.iter().map(|h| cell!(h)).collect();
-    for ref key in data.keys() {
-        row_header.push(cell!(&key));
-    }
-    table.add_row(Row::new(row_header));
-    table
 }
